Norme compliant

+++ srcs/sprite.c
@@ -6,7 +6,7 @@
 /*   By: alao <alao@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/28 12:06:05 by alao              #+#    #+#             */
-/*   Updated: 2016/12/01 18:50:52 by alao             ###   ########.fr       */
+/*   Updated: 2016/12/02 14:21:25 by alao             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,7 +45,7 @@ static void		wolf_sprite_renderer(t_wolf *w, int x, int y, t_sprite *s)
 		y = w->v.sp_ystart;
 		while (y < w->v.sp_yend)
 		{
-			tmp = (s->scaled) ?(y - (int)(256 / w->v.sp_transy)) : y;
+			tmp = (s->scaled) ? (y - (int)(256 / w->v.sp_transy)) : y;
 			tmp = tmp * IMG_SPR_Y - WIN_Y * 128 + w->v.sp_y * 128;
 			sp_texy = ((tmp * IMG_SPR_Y) / w->v.sp_y) / 256;
 			clr = wolf_mlx_get_color(w, sprite, sp_texx, sp_texy);
@@ -74,7 +74,7 @@ static void		wolf_sprite_calculator(t_wolf *w, t_sprite *s)
 	tmp = 1.0 / (w->p.pln_x * w->p.dir_y - w->p.dir_x * w->p.pln_y);
 	w->v.sp_transx = tmp * (w->p.dir_y * w->v.sp_x - w->p.dir_x * w->v.sp_y);
 	w->v.sp_transy = tmp * (-w->p.pln_y * w->v.sp_x + w->p.pln_x * w->v.sp_y);
-	w->v.sp_fact = (int)(WIN_CX * (1 + w->v.sp_transx / w->v.sp_transy));
+	w->v.sp_fact = WIN_CX * (1 + w->v.sp_transx / w->v.sp_transy);
 	w->v.sp_y = abs((int)(WIN_Y / w->v.sp_transy));
 	(s->scaled) ? w->v.sp_y /= SPR_SCALE : (0);
 	w->v.sp_ystart = -w->v.sp_y / 2 + WIN_CY;
